psychrag/requirements.txt
langchain
streamlit
langchain-groq
sentence_transformers
pypdf
faiss-cpu
langchain-community
python-dotenv
langchain-huggingface 
huggingface_hub